+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Env vars being set or unset for the app on deploy update errors
 - Typo on server pkg
+- [server] Return all deploy errors to the client
 
 ## [0.11.0] - 2017-12-11
 ### Added